feat: modify hook to returned combined fee and discount values
