+++ libs/fills/src/lib/fills.mock.ts
@@ -55,12 +55,18 @@ export const generateFill = (override?: PartialDeep<FillFieldsFragment>) => {
       makerFee: '100',
       infrastructureFee: '100',
       liquidityFee: '100',
+      treasuryFee: '0',
+      buyBackFee: '0',
+      highVolumeMakerFee: '0',
     },
     sellerFee: {
       __typename: 'TradeFee',
       makerFee: '200',
       infrastructureFee: '200',
       liquidityFee: '200',
+      treasuryFee: '0',
+      buyBackFee: '0',
+      highVolumeMakerFee: '0',
     },
     market: {
       __typename: 'Market',
@@ -96,12 +102,18 @@ export const fillFieldsFragment = (
       makerFee: '100',
       infrastructureFee: '100',
       liquidityFee: '100',
+      treasuryFee: '0',
+      buyBackFee: '0',
+      highVolumeMakerFee: '0',
     },
     sellerFee: {
       __typename: 'TradeFee',
       makerFee: '200',
       infrastructureFee: '200',
       liquidityFee: '200',
+      treasuryFee: '0',
+      buyBackFee: '0',
+      highVolumeMakerFee: '0',
     },
     market: {
       __typename: 'Market',
@@ -184,12 +196,18 @@ export const fillsEventSubscription = (
           makerFee: '100',
           infrastructureFee: '100',
           liquidityFee: '100',
+          treasuryFee: '0',
+          buyBackFee: '0',
+          highVolumeMakerFee: '0',
         },
         sellerFee: {
           __typename: 'TradeFee',
           makerFee: '200',
           infrastructureFee: '200',
           liquidityFee: '200',
+          treasuryFee: '0',
+          buyBackFee: '0',
+          highVolumeMakerFee: '0',
         },
       },
     ],
